requirements.txt
pandas==2.2.2
pyodbc==5.1.0
paramiko==3.4.0
requests==2.32.3
openpyxl==3.1.4
python-dotenv==1.0.1
schedule==1.2.2
